move unnecessary clone tests
